lang-app/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
openai==1.14.0
httpx==0.27.0
requests==2.31.0
pydantic==2.6.3